search/requirements.txt
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
azure-ai-vision-imageanalysis==1.0.0b2
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-search-documents==11.6.0b4
azure-storage-blob==12.20.0
blinker==1.8.2
boto3==1.34.133
botocore==1.34.133
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cohere==5.5.8
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.8
cycler==0.12.1
distro==1.9.0
docopt==0.6.2
exceptiongroup==1.2.1
fastavro==1.9.4
filelock==3.15.4
Flask==3.0.3
fonttools==4.53.0
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.4
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
msal==1.29.0
msal-extensions==1.2.0
num2words==0.5.13
numpy==2.0.0
openai==1.35.4
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
pillow==10.3.0
plotly==5.22.0
portalocker==2.10.0
protobuf==5.27.2
pyarrow==16.1.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
s3transfer==0.10.2
scikit-learn==1.5.0
scipy==1.14.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.36.0
tenacity==8.4.2
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
types-requests==2.32.0.20240622
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
waitress==3.0.0
watchdog==4.0.1
Werkzeug==3.0.3
Flask-Cors==4.0.1